Remove dead code and refactor namespaces
- Remove unused files
  - Delete DataReaderMapper.orig
  - Delete ReflectionExtensions.cs

- Refactor namespaces to be more in line with main AutoMapper library
