+++ AutoMapper.Data.Tests/DataReaderMapping.cs
@@ -4,9 +4,10 @@
     using System.Collections.Generic;
     using System.Data;
     using System.Linq;
+    using Configuration.Conventions;
     using Mappers;
-    using Xunit;
     using Should;
+    using Xunit;
 
     public class When_mapping_a_data_reader_to_a_dto
     {
