+++ AutoMapper.Data/Utils/DataReaderHelper.cs
@@ -1,14 +1,10 @@
 ﻿using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Data;
 using System.Linq;
-using System.Linq.Expressions;
 using System.Reflection;
-using AutoMapper.Data;
-using static System.Linq.Expressions.Expression;
 
-namespace AutoMapper.Utils
+namespace AutoMapper.Data.Utils
 {
     public static class DataReaderHelper
     {
